Initial solutions for day 7
Very slapdash hand eval function (but it works)

For part 2 just brute forced over all choices for J (but checked for 4
or 5 J's first). Thinking about it now there shouldn't ever be a reason
to make different Js into different cards though (only need that if we
have straights or something)...so that could have been a lot simpler.

Luckily the way I wrote part1 didn't need any change for the aspect
that J's are wild for hand type but not wild for the tiebreak.