Initial solutions for day 10
A bit of a twist to find the right starting point (have to consider
neighbors of S that "connect" back to it) but after that it is standard
BFS to find the loop & assign distances to it

In part 2 we need the loop so that part is all reused. For the "inside"
determination, imagine shooting rays left-to-right, each time we cross a
pipe we go switch between inside / outside. We also imagine shifting
things a bit vertically to avoid cases where the ray has infinite
intersections with a horizontal pipe: I chose to imagine the pipe always
lying along the bottom edge when moving horizontally so only
intersections with "|", "L", and "J" matter.

Wasted a bit of time due to missing the bit in the description that says
we should count even "junk" pipes inside the loop; I was originally
counting just "." tiles. The distance numbers make it easy to identify
which pipes are "junk" so that is not hard to deal with.