Initial solutions for day 11
number of galaxies is small enough that we don't need to do any fancy
scanline tricks. This problem can actually be decomposed to treat each
dimension separately which could be worth it if there are a superlinear
number of galaxies.

chose an appropriately flexible method for part1 that made part2 a
trivial change. however, I lost time because I impulsively changed the
adjustment to part2 to 1M instead of (1M-1) (had to wait an extra minute
due to that)

consider some updates to the template ("parse" is a bit of a misnomer
for the kind of preprocessing I wanted to do here, also is useful to
pass in the different parts as separate args)