Initial solutions for day 12
"quick" solution for part 1 ended up wasting some time because I ended
up grafting on the "needs all groups" case afterwards. Would probably
have been best to just do the dynamic programming approach from the
start. Also note that this DP approach never deals with partial groups
which simplifies things.

Only trick for that is to add an extra "." at the end of the input
because each step requires a "." at the end of the group (to avoid
making two groups directly next to each other). I also got the recursive
indexing wrong-by-one at first for that detail.

Lost even more time due to incorrectly early-returning when there is no
way to make the group (answer is decided if the char is "#" but if it is
"?" we still have the option of not making a group...this would have
been less error-prone if I had just split the three cases explicitly).

Finally I had a mistake in the base-case where when all groups are
finished I was checking that the record was all "." instead of checking
for no "#".

Luckily the examples helped catch each of these problems without wasting
too much time debugging.