Initial solutions for day 14
LATE START today TT

Simple "just do it" approach for the rolling: scan in order dropping
rocks off where they belong, "resetting" that pointer when we see a #

In part 1 we don't need to actually move anything to compute the answer.
I definitely could predict that part 2 would involve rolling in the
other directions, though I'm not sure that I would have changed much to
handle that. Just wrote a similar loop with some indicies / directions
flipped for each one.

For part 2 rely on the assumption that it won't take long before we get
a cycle. I'm not sure how to prove that exactly but it certainly holds
for the data (repeats within a couple hundred spins). So then it's a
straightforward process to find the length of that cycle and mod down
the 1B goal to a value that we can iterate to.