Initial solutions for day 17
Solved using a priority queue. I also added pruning based on revisiting
a state (position + direction) with a worse value. This was good enough.

I wasted some time on part 1 because I was only preventing going
backwards, and not preventing continuing on in the same direction (silly
oversight).

The adjustment for part 2 was easy as it only affects which are the
possible next states and no other logic needs to change.