Initial solutions for day 19
Wish I'd written a simple parser util lib. Spent a bunch of time on the
ad-hoc parsing. Anyway this will be a good example to use to develop
that.

Ran into some issues with python lambda var scope. Not quite sure what
was going wrong there exactly but I eventually just abandoned the
approach.

For part 2 we just split the "part" up into ranges based on where the
decision points change. The way I did it was sort of ugly / lots of
code...it probably would have been fine to just skip all of that and
simply *try* all the values in the range to see where they go since we
know each condition only needs to check at most 4000 values.