Initial solutions for day 20
Simulation aspect was straightforward to implement (nothing tricky
required).

Part 2 requires detailed manual analysis of the actual structure of the
input. It consists of 4 independent counters that all need to align.

https://csacademy.com/app/graph_editor/ was quite useful to help sort
out what the graph structure looks like.

In refinement we can partially automate this analysis. The key point is
we could focus on the 4 "test/reset" conjunction nodes and watch for
when they flip; it's easy to identify which nodes these are in the input
as they are the conjunction nodes with a bunch of outputs.

I did try something like that at first but I was only looking at the
state after everything settled, which wasn't enough because the node
states reset by then: we need to watch for pulse sent from the reset
node *during* the settling process.