Initial solutions for day 21
Part 1 is straightforward. Note that we don't need to worry too much
about the "exact distance" constraint: if the minimum distance to the
plot is z, then we can also reach it after any z + 2k number of steps,
and also cannot reach it with any other numbers of steps (parity). So we
can derive the answer based on minimum distance values.

For part 2, the idea is that if we are considering a plot in a distant
enough "copy", the minimum distance to that plot is the same as the
minimum distance to that plot in a nearby copy plus a multiple of the
map's side length. I expanded the map to a (2k+1)*(2k+1) version, and
computed the true distances in that. Then, took the plots in the
"boundary cells" of that map as the "nearby" exemplars to determine how
many distant copies would also be reachable in that direction: I
expanded most boundary cells in just one cardinal direction and then had
a special case for the corners.

The parity point noted above also comes into play because the map's side
length is actually odd; so only ~half of the distant copies of any given
plot will actually be reachable in the desired number of steps.

The computed answer was the same for 2k+1 = 3, 5, 7 for this input.
Getting the right answer for the example case @ 5000 steps required 2k+1
= 7.