Initial solutions for day 22
The dimensions are tiny so it is easy to sort out the brick
relationships by building up a height map on X-Y.

For part 1 we just keep track of which bricks were the "sole supporters"
for any brick that came later.

For part 2 we leverage similar information but now need to know
transitively which bricks rely on the removed brick. There is probably a
more efficient way to do this graph-theoretically but I did the simple
approach of just checking whether there is still a path to ground
("None") in the graph when I ignore the given brick. This op can be
memoized to compute it efficiently enough (the memo resets for each
choice of deleted brick).