Initial solutions for day 23
Fairly hacky. The key point is that the provided map is mostly
single-choice trails that link up at a few intersection points, and also
(for part 1) those intersection points have slopes.

For part 1 assumed that it was ok to memoize at the decision points.

For part 2 this definitely won't work. This should be solved by reducing
the graph down into a few nodes connected by weighted edges (but
ultimately some kind of brute-force is required?). However in practice I
just let the ultra-simple brute-force go and it found the right answer
after about 20m (it didn't terminate then, but I just tried submitting
the best so far, takes about 70m to terminate).

There *might* be something clever we could do based on cutting cycles
(depending on what the cycle structure looks like) and/or somehow
applying eulerian paths to a subgraph etc.