Initial solutions for day 24
Part 1 is trivial but I overcomplicated it because I expected part 2 to
involve using the z dimension.

My geom intersection notes came in useful here. Worth revisiting (might
be able to simplify some things using triple-scalar-product properties?)

For part 2 we can systematically leverage the existence of a solution to
narrow things down. Refer to working.notes for details of how that works.