Solution for Day 25
Hammer it using min-cut-max-flow.

I didn't realize there wouldn't be a part 2, so I kind of took it slow
thinking that I'd probably spend a lot more time on part 2 anyway.

Heavily followed an old flow solution I had lying around (but that was
slightly different because it was C++ and also had directed edges with
varying capacity: this flow is simpler because the edges are
bidirectional and the flow is 0-1.

Since we don't know what the cut separates, try various sinks until we
get a cut of 3. Stoer-Wagner algorithm could be used to do something
smarter (but that is so much extra work here...)

There's also probabilistic approach to global min-cut: Karger's
algorithm. Since we even know the cut-size = 3 we can probably do even
better than usual with this.