Solution for day 2
Quite fast (but there were issues with the site this time!)

zip(lst, tail) is a handy idiom for this kind of thing.

For part 2 just did the most naive approach; in theory one could try to
be cleverer but the reports were very short so no reason to do that.