Solution for day 3
Got to use my fancy parsing lib. The parsing in this case boils down to
repeating "check for real instruction, or skip a char", which could be
done easily enough using plain regex (but the lib provides a nice
interface for dealing with extracting results from regex).

Was a bit tricked by the example data which is different between parts 1
and 2; when I tried my part 2 solution on the example from part 1 I got
a different answer than expected. Should be careful to watch for this
(big clue is that the example was explicitly written out in part 2).