[2024] Solution for day 5
Originally I just used bubblesort, but built-in sort can be used too.

The input is described as a "very specific order" which actually means
there is an entry in the rules for every possible pair. In other words
we don't need to the transitive closure of the relationships (which
bubblesort would do implicitly, if written in an appropriately tentative
way).

Anyway the outline of the solution is simple: sort each list based on
the ordering and add the middle number to the right total depending on
if it changed or not.