[2024] Solution for day 6
This is fairly raw (not modified much from my advent-time code).

Just simulate directly: for part1 stop when we leave the grid, for part2
add cycle detection. One minor optimization is that for part2 we only
need to try putting obstacles on the spots visited in part1.

Should be cleaned up / sped up a bit. Ideas:
- refactor the traversal part to a separate generator that just yields
  the steps on the path
- preprocess the input to get the next "waypoint" in each direction,
  then traversal can just output those instead of every point
- (unsure if worthwhile) start the guard in front of the new obstacle
  instead of from the start point