[2025] Solution for day 3
Dynamic programming based on the idea that whenever we pick the last
digit, the digits before that form a smaller subproblem (1 fewer digit +
a prefix of the bank). At each digit the choice is simply whether to
pick it or not.

Got a bit slowed down messing with the table indices (also was
indecisive about whether to do this backwards or forwards). A top-down
memoizing approach might have been less error-prone to write.