[2025] Improved solution for day 4
Simple tweak to only look at neighbors of last-peeled layer when
computing the next one...since we already identify all those positions
when evaluating the "peelability" might as well reuse the information.

Structuring this as "layers" is a bit unnatural (just doing one removal
at a time would be a bit simpler code-wise) but I'm doing it this way
due to what part1 is asking.

I couldn't come up with a substantially different approach for this
problem, and also suspect that the best way to further speed this up
would be to use a more performant array (e.g. something from numpy).