[2025] Solution for day 9
Ended up using coordinate compression, but for whatever reason I forgot
how to avoid typical mistakes with this kind of question (partially I
confused myself alternating between thinking of grid squares vs. integer
points). A sweep-based solution (without explicitly compressing)
probably would have been quicker to get right.

Conceptually the idea here is to look at rays slightly above/right of
the integer points to avoid the corner (literally) cases.

Originally I built a 2D boolean map of the compressed "world" and
scanned that for rectangle "fullness", which wasted a lot of unnecessary
time (both coding and runtime). Once I had the cross points for each row
a simple check that there aren't any cross points between the left and
right sides of the rectangle (and that we have an odd number before the
first one) is all that is needed and also simpler to implement.

Note: for my data set I found it was faster to switch x/y.