[2025] Add faster solution for day 8
Improve the edge-finding logic to find the shorter edges quickly. This
is ultimately just a heuristic approach and could end up doing more work
(in fact the generator doesn't even know how to stop).

The idea is fairly simple, just maintain a sliding window based on
x-coordinate that cuts off at the highest length we're interested in,
and iterate by increasing this bound (doubling each time).

Note that when we find edges that are longer than that, we need to
ignore them, because we aren't able to guarantee they'll be in the right
sorted order yet. And we also use a lower bound to avoid repeating any
edges that we would have found in a previous phase.