[2025] Solution for day 10
Based on the observation from
https://www.reddit.com/r/adventofcode/comments/1pk87hl/2025_day_10_part_2_bifurcate_your_way_to_victory/
which I shall summarize as:

For any set of button presses achieving a given joltage vector, some
subset of the buttons are pressed an odd number of times, and the others
an even number of times. If we press those odd buttons once each, the
resulting joltages *must* be equal modulo 2 to the final total. We can
use this fact to prune the possibilities for the odd subset. And then,
the rest of the problem can be divided in half.

Speed here is mostly affected by the time to precompute the properties
of the subsets. One idea for speeding that up is to generate them in
Gray code order (thus only needing to update state of one button each
time). I presume using numpy would also make that faster since it is
just adding up a bunch of 0-1 vectors.