[2025] Z3-based solution for day 10
This is slower than parity-based search (at least for the sizes involved
in our data), but useful as a simple demo of how to use z3 for such
problems.

I also looked at implementing simplex but in this case we will run into
issues staying in integer domain. Apparently a branch-and-bound approach
can work. I also wondered if recasting the problem as a 0/1 selection
(bundling multiple button presses into groups of 1,2,4,8...) could help
with that.